SCRUM-226 fix implementation + new file for timer
