+++ app/routers/sio_game.py
@@ -55,6 +55,4 @@ async def connect(sid, environ, auth):
 
         await game_events.emit_turn_info(game_id, db)
 
-        asyncio.create_task(game_events.emit_timer(game_id, player.id, db))
-
         await game_events.emit_opponents_total_mov_cards(game_id, db)
