+++ app/services/game_events.py
@@ -6,8 +6,7 @@ from app.schemas.player import PlayerResponseSchema, WinnerSchema
 from app.services.cards import fetch_figure_cards, fetch_movement_cards
 from app.services.board import get_board
 from app.services.figures import figures_event
-from app.services.game import end_turn
-
+from app.services.timer import handle_timer
 
 async def disconnect_player_socket(player_id, game_id):
     broadcast = Broadcast()
@@ -53,23 +52,8 @@ async def emit_turn_info(game_id, db):
     # send the turn info to all players in the lobby
     await broadcast.broadcast(sio.sio_game, game_id, "turn", turn_info)
 
-    # Start the timer for the new turn
-    asyncio.create_task(emit_timer(game_id, player.id, db))
-
-
-async def emit_timer(game_id, player_id, db):
-
-    broadcast = Broadcast()
-    time_left = 120
-
-    while time_left > 0:
-        await broadcast.broadcast(
-            sio.sio_game, game_id, "timer", {"time": time_left}
-        )
-        await asyncio.sleep(1)
-        time_left -= 1
-
-    end_turn(game_id, player_id, db)
+    # start the timer for the current player
+    await handle_timer(game_id, player.id, db)
 
 
 async def emit_winner(game_id, winner_id, db):
